Rename DrawableOsuContextMenuItem -> DrawableOsuMenuItem, and move to OsuMenu.
